+++ Project.toml
@@ -2,7 +2,7 @@ name = "FractionalCalculus"
 uuid = "638fb199-4bb2-4014-80c8-6dc0d90f156b"
 license = "MIT"
 authors = ["Qingyu Qu <erikqqy123@gmail.com>"]
-version = "0.3.0"
+version = "0.4.0"
 
 [deps]
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
@@ -11,20 +11,27 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 SpecialMatrices = "928aab9d-ef52-54ac-8ca1-acd7ca42c160"
-SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
-Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 
 [compat]
 ForwardDiff = "0.10.36"
 InvertedIndices = "1.3"
+LinearAlgebra = "1.10"
 QuadGK = "2.9"
 SpecialFunctions = "2.3"
 SpecialMatrices = "3"
 SymbolicUtils = "1.5"
 Symbolics = "5.13"
+Test = "1.10"
 UnPack = "1.0.2"
-julia = "1.9"
+julia = "1.10"
+
+[weakdeps]
+SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+
+[extensions]
+FractionalCalculusSymbolicsExt = ["Symbolics", "SymbolicUtils"]
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
