+++ test/aqua.jl
@@ -1,0 +1,7 @@
+using Aqua
+using Test
+using FractionalCalculus
+
+@testset "Aqua" begin
+    Aqua.test_all(FractionalCalculus)
+end
