+++ test/runtests.jl
@@ -4,6 +4,7 @@ using SpecialFunctions
 using Test
 
 @testset "FractionalCalculus" begin
+    include("aqua.jl")
     include("./Derivative.jl")
     include("./Integral.jl")
     include("./SymbolicDiffTests.jl")
